Speed Handler Module
A module that specifically handles the movement speed of players. Added to make speed simpler to modify, as previously it was all over the place where I would be modifying speed in other scripts as needed. This is the only script where speed modification should occur in order to make things easier.