Status Module and Stun Module
Added a revised Status Effect system that is intended to be modular, and work with other modules that contain function for how certain status effects work. The first status is a Stun rework. This change has overall made the system feel more smooth, and makes it easier to work with in other scripts.