Reworked m1Handler
Reworked the way m1s are handled, making it so the hitbox for the attack is now timed with the exact frame the attack happens in the animation. This change has made combat feel way smoother. The swing cooldown is also changed to happen on the 20th frame of an animation.