New Character Data & Attribute
Added data for a new character, wDon. Additionally, characters now have an "amToWait" attribute. This is used to shoot a hitbox out during the desired exact time of a character's animation. To be specific, the math is simply: [Frame to shoot hitbox] / [Total frames]. So, using wDon as an example, the hitbox should shoot out at frame 15 of her 60 frame attack animations (15/60).