Client VFX modules.
Some old client VFX modules I made at the beginning of the semester. I consider oldClientVFXModule deprecated, but decided to leave it in the repo in case I'd want to reference it in the future. These scripts simply manage visual effects playing on the client (as having them handled on the server is laggy).